eb524f4 COMP: Avoid compiler warnings with parital specialization
14bef3f COMP: Code specific to 3 dims was specialized
b6b6b12 COMP: Removed unreachable code.
480500d COMP: Check existance of variable is defined.
cbd07da COMP:  Change Superclass to real Superclass.
d7b6427 COMP: Explicitly set public/static/const/virtual
7ea74fb COMP: Remove external template designation from VNL
8fd6ee5 COMP: Fix incompatible format string conversion
bd8d107 COMP: Force explicit integer conversion
6fed567 COMP:  Intel compiler option settings.

+++ Modules/Filtering/FFT/test/itkRealFFTTest.h
@@ -57,7 +57,7 @@ test_fft(unsigned int *SizeOfDimensions)
   for (unsigned int i = 0; i < VImageDimensions; i++)
     {
     imageSize.SetElement( i, SizeOfDimensions[i] );
-    imageIndex.SetElement( i, -2*i - 5 ); // Test for handling non-zero
+    imageIndex.SetElement( i, -2*static_cast<int>(i) - 5 ); // Test for handling non-zero
                                           // image indices correctly
     }
